Remove HttpBuilder
