+++ query-engine/query-engine/src/main.rs
@@ -1,13 +1,12 @@
 #[macro_use]
 extern crate tracing;
 
-use cli::*;
-use error::*;
+use cli::CliCommand;
+use error::PrismaError;
 use once_cell::sync::Lazy;
-use opt::*;
+use opt::PrismaOpt;
 use request_handlers::{PrismaRequest, PrismaResponse, RequestHandler};
-use server::{HttpServer, HttpServerBuilder};
-use std::{convert::TryFrom, error::Error, net::SocketAddr, process};
+use std::{convert::TryFrom, error::Error, process};
 use structopt::StructOpt;
 use tracing::subscriber;
 use tracing_subscriber::{EnvFilter, FmtSubscriber};
@@ -54,34 +53,7 @@ async fn main() -> Result<(), AnyError> {
         }
         Err(PrismaError::InvocationError(_)) => {
             set_panic_hook()?;
-            let ip = opts.host.parse().expect("Host was not a valid IP address");
-            let addr = SocketAddr::new(ip, opts.port);
-
-            eprintln!("Printing to stderr for debugging");
-            eprintln!("Listening on {}:{}", opts.host, opts.port);
-
-            let create_builder = move || {
-                let config = opts.configuration(false)?;
-                let datamodel = opts.datamodel(false)?;
-
-                PrismaResult::<HttpServerBuilder>::Ok(
-                    HttpServer::builder(addr, config, datamodel)
-                        .legacy(opts.legacy)
-                        .enable_raw_queries(opts.enable_raw_queries)
-                        .enable_playground(opts.enable_playground),
-                )
-            };
-
-            let builder = match create_builder() {
-                Err(err) => {
-                    info!("Encountered error during initialization:");
-                    err.render_as_json().expect("error rendering");
-                    process::exit(1);
-                }
-                Ok(builder) => builder,
-            };
-
-            if let Err(err) = builder.build().await {
+            if let Err(err) = server::listen(opts).await {
                 info!("Encountered error during initialization:");
                 err.render_as_json().expect("error rendering");
                 process::exit(1);
